+++ NEWS.md
@@ -1,3 +1,19 @@
+# LOSYkolory 0.4.0 (17.09.2025)
+
+## Aktualizacje
+
+-   Wprowadzono zmiany pozwalające na działanie z pakietem *ggplot2* w wersji 4.0.0:
+
+    -   `zmien_wielkosc_czcionek()` nie próbuje modyfikować definicji wielkości czcionki w *warstwach* tekstowych wykresu, które mają domyślną wartość `size` zdefiniowaną jako wywołanie `from_theme(fontsize)` (w praktyce: zwracają `TRUE` przy wywołaniu na nich `is.language()`), za to modyfikuje wielkość czcionki w domyślnych parametrach *geometrii* zdefiniowanych w ramach *tematu* wykresu (które propagują się na domyśle wartości *estetyki* `size` definiowane w ramach poszczególnych *warstw* tekstowych),
+    -   W związku z tym w `zmien_wielkosc_czcionek()` zmieniono nazwę argumentu `textGeoms` na `geoms`, aby zaznaczyć, że nominalnie zmiana może też dotyczyć innych *warstw* (jeśli tylko w odniesieniu do jakiejś *estetyki* jest w nich wykorzystywane mapowanie postaci `from_theme(fontsize)`).
+    -   W środowisku z pakietem *ggplot2* w wersji 4.0.0 lub wyższej wywołanie `theme_losy()` z podanym argumentem `base_family` modyfikuje także zapisywaną w *temacie* wykresu domyślną wartość *estetyki* `family`, która z kolei propaguje się na wszystkie geometrie, w których *estetyka* `family` nie została zdefiniowana explicite (czy to przez zadanie konkretnej wartości, czy to przez przypisanie mapowania na zmienną),
+        -   W związku z tym w typowych zastosowaniach **nie** trzeba już będzie korzystać oddzielnie z funkcji `zmien_rodzine_czcionek()`,
+    -   Podniesiono zależność od pakietu *ggplot2* do wersji 4.0.0, aby uniknąć problemów ze zróżnicowanymi wynikami działania funkcji w zależności od środowiska, w którym są uruchamiane.
+
+## Usunięte błędy
+
+-   W `theme_losy()` i `zmien_rodzine_czcionek()` przy weryfikacji, czy podana rodzina czcionek jest dostępna w systemie uwzględniane są również wartości zawsze obsługiwane przez pakiet *ggplot2*, tj. "sans", "serif" i "mono".
+
 # LOSYkolory 0.3.0 (04.08.2025)
 
 ## Nowe funkcje i funkcjonalności
@@ -46,7 +62,7 @@
 -   `palette_losy()` obsługuje nowe wartości argumentu `type`:
     -   "bezrobocie" - do wykresów prezentujących odsetek bezrobotnych wśród absolwentów,
     -   "formyZatrudn" - do wykresów prezentujących formy zatrudnienia w raportach szkolnych,
-    -   "grupaOdn" - do wykresów w raportach szkolnych, w których kolor odróżnia daną szkołę od odpowiadajacej jej *grupy odniesienia*,
+    -   "grupaOdn" - do wykresów w raportach szkolnych, w których kolor odróżnia daną szkołę od odpowiadającej jej *grupy odniesienia*,
     -   "kontMlodocPracEdu" - do wykresów opisujących kontynuowanie (lub nie) zatrudnienia u pracodawcy zatrudniającego wcześniej absolwenta jako młodocianego pracownika (w raportach szkolnych),
     -   "kontMlodocPracNedu" - do wykresów opisujących kontynuowanie (lub nie) zatrudnienia u pracodawcy zatrudniającego wcześniej absolwenta jako młodocianego pracownika (w raportach szkolnych),
     -   "okresPracy" - do wykresów prezentujących, jak dużą część monitorowanego okresu absolwenci byli zatrudnieni (w raportach szkolnych),